Add subtasks feature implementation
This feature allows to create subtasks on every task. The subtasks are displayed as points and settings can be adjusted in the respective Tasks tab. It allows to adjust the position in the timeline squares (inline, outside or online). It is possible to adjust the border size and color. This settings are stored in the persistent configuration files of the user computer.